Remove some more dependencies on FieldAttribute
